+++ packages/components/src/index.js
@@ -18,6 +18,7 @@ export { default as DatePicker } from './calendar/date-picker';
 export { default as DropdownButton } from './dropdown-button';
 export { default as EllipsisMenu } from './ellipsis-menu';
 export { default as EmptyContent } from './empty-content';
+export { default as ErrorBoundary } from './error-boundary';
 export { default as Flag } from './flag';
 export { default as Form } from './form';
 export { default as FilterPicker } from './filter-picker';